requirements.txt
Flask
langchain
ollama
python-dotenv
SQLAlchemy
Jinja2